Fixed cyclic dependency
