+++ x/upgrade/abci.go
@@ -3,8 +3,8 @@ package upgrade
 import (
 	"fmt"
 
-	abci "github.com/tendermint/tendermint/abci/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	abci "github.com/tendermint/tendermint/abci/types"
 )
 
 // BeginBlock will check if there is a scheduled plan and if it is ready to be executed.
