+++ x/upgrade/alias.go
@@ -34,6 +34,8 @@ var (
 	NewQueryAppliedParams            = types.NewQueryAppliedParams
 	NewKeeper                        = keeper.NewKeeper
 	NewQuerier                       = keeper.NewQuerier
+	UpgradeableStoreLoader           = types.UpgradeableStoreLoader
+	StoreLoaderWithUpgrade           = types.StoreLoaderWithUpgrade
 )
 
 type (
